PALMAS: Resolve build warning on Palmas
Resoved warning:
drivers/mfd/palmas-resource.c: In function 'palmas_enable_clk32kgaudio':
drivers/mfd/palmas-resource.c:63: warning: passing argument 3 of 'palmas_resource_read' from incompatible pointer type
drivers/mfd/palmas-resource.c:33: note: expected 'unsigned int *' but argument is of type 'u8 *'

+++ drivers/mfd/palmas-resource.c
@@ -57,7 +57,7 @@ static int palmas_resource_write(struct palmas *palmas, unsigned int reg,
 int palmas_enable_clk32kgaudio(struct palmas_resource *resource)
 {
 	int ret;
-	u8 reg;
+	unsigned int reg;
 
 	ret = palmas_resource_read(resource->palmas,
 			PALMAS_CLK32KGAUDIO_CTRL, &reg);
